# docs/swagger.yaml
basePath: /
definitions:
  models.ErrorResponse:
    properties:
      error:
        example: Error message here
        type: string
    type: object
  models.Food:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        example: Delicious dish
        type: string
      food_image:
        example: https://example.com/images/chicken.jpg
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      image_base64:
        description: Base64 data for photo uploads
        type: string
      image_file_id:
        example: file_id_here
        type: string
      menu_id:
        example: 507f1f77bcf86cd799439011
        type: string
      name:
        example: Grilled Chicken
        maxLength: 100
        minLength: 2
        type: string
      price:
        example: 150
        type: number
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    required:
    - food_image
    - menu_id
    - name
    - price
    type: object
  models.FoodResponse:
    properties:
      food:
        $ref: '#/definitions/models.Food'
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      message:
        example: Food created successfully
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@restaurant.com
    name: API Support
  description: This is a restaurant management system API with user authentication,
    food ordering, and invoice management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Restaurant Management API
  version: "1.0"
paths:
...
  /foods:
    get:
      consumes:
      - application/json
      description: Retrieve a complete list of all available food items in the restaurant
      produces:
      - application/json
      responses:
        "200":
          description: Array of all food items with details
          schema:
            items:
              $ref: '#/definitions/models.FoodResponse'
            type: array
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Database error while fetching foods
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Foods
      tags:
      - Food
    post:
      consumes:
      - multipart/form-data
      description: Create a new food item in the restaurant menu (Admin only)
      parameters:
      - description: Food Name
        in: formData
        name: name
        required: true
        type: string
      - description: Food Price
        in: formData
        name: price
        required: true
        type: number
      - description: Menu ID
        in: formData
        name: menu_id
        required: true
        type: string
      - description: Food Description
        in: formData
        name: description
        type: string
      - description: Food Image File
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Food created successfully with generated ID
          schema:
            $ref: '#/definitions/models.FoodResponse'
        "400":
          description: Invalid request body or validation failed
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Database error while creating food
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Food
      tags:
      - Food
  /foods/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a food item from the menu (Admin only)
      parameters:
      - description: Food MongoDB ObjectID
        example: '"507f1f77bcf86cd799439011"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Food deleted successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid MongoDB ObjectID format
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Food item not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Database error while deleting food
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Food
      tags:
      - Food
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a specific food item using
        its unique ID
      parameters:
      - description: Food MongoDB ObjectID
        example: '"507f1f77bcf86cd799439011"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Food item details
          schema:
            $ref: '#/definitions/models.FoodResponse'
        "400":
          description: Invalid MongoDB ObjectID format
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Food item not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Food by ID
      tags:
      - Food
    put:
      consumes:
      - multipart/form-data
      description: Update an existing food item's information (Admin only)
      parameters:
      - description: Food MongoDB ObjectID
        example: '"507f1f77bcf86cd799439011"'
        in: path
        name: id
        required: true
        type: string
      - description: Food Name
        in: formData
        name: name
        required: true
        type: string
      - description: Food Price
        in: formData
        name: price
        required: true
        type: number
      - description: Menu ID
        in: formData
        name: menu_id
        required: true
        type: string
      - description: Food Description
        in: formData
        name: description
        type: string
      - description: Food Image File
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Food updated successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid ID format or request body
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Food item not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Database error while updating food
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Food
      tags:
      - Food
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"